Revert "perf(rwc): precompute gist highlights during build"
This reverts commit 518d0fd16887f9ace9c5de8cb95cfb74e8434ce5.
