+++ package.json
@@ -10,13 +10,12 @@
 		"dev:content": "velite --config './.config/velite.config.ts' --watch",
 		"build:content": "velite --config './.config/velite.config.ts' --clean",
 		"generate-feeds": "tsx scripts/generate-feeds.ts",
-		"dev": "pnpm run generate-rwc && pnpm run dev:content & vite dev",
-		"generate-rwc": "tsx scripts/generate-rwc.ts",
-		"prebuild": "pnpm run build:content && pnpm run generate-rwc && pnpm run generate-feeds",
+		"dev": "pnpm run dev:content & vite dev",
+		"prebuild": "pnpm run build:content && pnpm run generate-feeds",
 		"build": "vite build",
 		"preview": "vite preview",
 		"compile": "tsc --watch",
-		"pretypecheck": "pnpm run build:content && pnpm run generate-rwc",
+		"pretypecheck": "pnpm run build:content",
 		"typecheck": "tsgo -p tsconfig.json --noEmit",
 		"cf-typegen": "wrangler types --env-interface CloudflareEnv ./src/cloudflare-env.d.ts",
 		"db:generate": "drizzle-kit generate",
